dom/imptests/webapps/DOMCore/tests/submissions/Ms2ger/test_Range-intersectsNode.html

<!doctype html>
<title>Range.intersectsNode</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
test(function() {
  var r = document.createRange();
  assert_throws(new TypeError(), function() { r.intersectsNode(null); });
  r.detach();
  assert_throws(new TypeError(), function() { r.intersectsNode(null); });
  assert_throws("INVALID_STATE_ERR", function() { r.intersectsNode(document); });
})
test(function() { // Step 2
  var r = document.createRange();
  assert_throws("INVALID_NODE_TYPE_ERR",
                function() { r.intersectsNode(document); });
  if (document.createAttribute) {
    assert_throws("INVALID_NODE_TYPE_ERR",
                  function() { r.intersectsNode(document.createAttribute("foo")); });
  }
})
test(function() { // Step 3
  var r = document.createRange();
  assert_throws("NOT_FOUND_ERR",
                function() { r.intersectsNode(document.createElement("div")); });
})
</script>
